requirements.txt
django==3.2.8
djangorestframework==3.11
requests==2.26.0
django-cors-headers==3.10.0
